all functions are now closures
