--- crates/error/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "error"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-anyhow = "1.0.66"
-thiserror = "1.0.37"
