+++ crates/one-liner/Cargo.toml
@@ -5,6 +5,5 @@ edition = "2021"
 
 [dependencies]
 report = { path = "../report" }
-error = { path = "../error" }
 parser = { path = "../parser" }
 crossterm = "0.25.0"
