+++ crates/snowc-errors/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "snowc-errors"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+# anyhow = "1.0.66"
+# thiserror = "1.0.37"
