+++ crates/snowc-errors/src/lib.rs
