+++ samples/changing_precedence.snow
@@ -1,0 +1,1 @@
+fn main = (((1)));
